+++ rt/__init__.py
@@ -1,6 +1,6 @@
-from .data import RTData, RTMultipartData
-from .exc import *
-from .frontend import RTFrontEnd
-from .interface import RTInterface
-from .response import RTResponse
-from .session import RTSession
+from .data import RTData, RTMultipartData  # noqa
+from .exc import *  # noqa
+from .frontend import RTFrontEnd  # noqa
+from .interface import RTInterface  # noqa
+from .response import RTResponse  # noqa
+from .session import RTSession  # noqa